+++ core/learning/__init__.py
@@ -1,0 +1,1 @@
+"""Compass learning package."""